refactor - arrange codes to their role in InputView and OutputView
