+++ src/main/java/bowling/view/InputView.java
@@ -1,6 +1,8 @@
 package bowling.view;
 
 public interface InputView {
+    void questionPlayerName();
     String playerName();
+    void questionPinNumber(int frame);
     int pinNumber();
 }
