+++ src/main/java/bowling/view/OutputView.java
@@ -3,7 +3,6 @@ package bowling.view;
 import bowling.domain.Frames;
 
 public interface OutputView {
-    void questionPlayerName();
-    void questionPinNumber(int frame);
+
     void resultsOfBowling(Frames frames);
 }
